Include Guzzle manually
This should  fix the  plugin when  installed  globally,  It's a terribly
nasty hack. But it's the only thing that will work when Composer doesn't
get it's crap together  and actually gives us a callback or exposes some
API we can call for autoloading.

The problem here lies in  the fact that Composer loads extra files (such
as a classmap) post-plugin. so when we are loaded, the function files of
Guzzle aren't loaded.  This can be done by  manually  calling require on
those files.

Luckily Guzzle already  has a system in place to prevent  double-loading
their functions and we can safely include it.

Github issues:
Fix Elendev/nexus-composer-push#11

Signed-off-by: Ted van Riel <tvanriel@wendrich.com>